Make Merkle tree depth variable consistant (d-->m)
This fixes an issue where in the doc sometimes we use d and sometimes we use m to refer to the Merkle tree depth. We should always use m because that is what BIP-341 uses

This commit also makes a minor change of ensure math expresses are inside double single quotes.